requirements.txt
asgiref==3.9.1
Django==5.2.6
django-phonenumber-field==8.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
phonenumbers==9.0.15
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
six==1.17.0
slugify==0.0.1
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
xlrd==2.0.2
